New launcher and icon images, added comments to the latest updated files and adjusted some files for better clarity.
