+++ SnakeBite/App.config
@@ -18,10 +18,10 @@
         <value>False</value>
       </setting>
       <setting name="EnableSound" serializeAs="String">
-        <value>True</value>
+        <value>False</value>
       </setting>
       <setting name="ThemeFile" serializeAs="String">
-        <value/>
+        <value />
       </setting>
       <setting name="LastSBVersion" serializeAs="String">
         <value>0.8.2.0</value>
