+++ SnakeBite/Properties/Settings.settings
@@ -9,7 +9,7 @@
       <Value Profile="(Default)">False</Value>
     </Setting>
     <Setting Name="EnableSound" Type="System.Boolean" Scope="User">
-      <Value Profile="(Default)">True</Value>
+      <Value Profile="(Default)">False</Value>
     </Setting>
     <Setting Name="ThemeFile" Type="System.String" Scope="User">
       <Value Profile="(Default)" />
